requirements.txt
aiohttp==3.8.1

aiosignal==1.2.0

async-timeout==4.0.2

attrs==21.4.0

certifi==2022.6.15

charset-normalizer==2.0.12

dataclasses-json==0.5.7

git+https://github.com/Rapptz/discord.py@13359372887abced96566bbf26a96e065faee782#egg=discord-py

git+https://github.com/Rapptz/discord.py#egg=discord.py

frozenlist==1.3.0

idna==3.3

isodate==0.6.1

marshmallow==3.16.0

marshmallow-enum==1.5.1

motor==3.0.0

multidict==6.0.2

mypy-extensions==0.4.3

oauthlib==3.2.0

packaging==21.3

pymongo==4.1.1

pyparsing==3.0.9

python-dotenv==0.20.0

python-youtube==0.8.2

requests==2.28.0

requests-oauthlib==1.3.1

six==1.16.0

typing-extensions==4.2.0

typing-inspect==0.7.1

urllib3==1.26.9

yarl==1.7.2

